+++ src/extensions/softgfx/wg_linearbackend.h
@@ -78,7 +78,7 @@ namespace wg
 		void	setCanvas( CanvasRef ref ) override;
 
 
-		void	processCommands( const int32_t* pBeg, const int32_t * pEnd) override;
+		void	processCommands( const uint16_t* pBeg, const uint16_t * pEnd) override;
 
 
 	protected:
